requirements_cloud.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
easyocr
torch
torchvision
numpy
